temporarily active users don't have to do recaptcha again
