+++ client/src/components/child-components/ImageView.js
@@ -1,0 +1,10 @@
+export const ImageView = (props) => {
+    const image = props.image;
+    console.log("image:", image)
+    return (
+      <section className='image-view'>
+        <img src = {image.image_url} alt={image.image_caption}/>
+        <h4>{image.image_caption}</h4>
+      </section>
+    )
+}
