+++ client/src/components/child-components/MemoryTile.js
@@ -1,9 +1,9 @@
-import { React, useState, useRef, useEffect} from 'react'
-import {v4 as uuidv4} from 'uuid'
+import { React, useState, useEffect} from 'react'
 import { handleGet, handlePost } from '../../services/requests-service';
 import axios from 'axios';
-import ReCAPTCHA from "react-google-recaptcha"
 import "../../styles/memories.css"
+import { CommentsView } from './CommentsView';
+import { ImageView } from './ImageView';
 const NODE_URL = process.env.REACT_APP_NODE_LOCAL || process.env.REACT_APP_NODE_PROD
 //child component for each memory
 export const MemoryTile = (props) => {
@@ -15,7 +15,7 @@ export const MemoryTile = (props) => {
   const [showComments, setShowComments] = useState(false); //show and hide comment view
   const [showImage, setShowImage] = useState(false); //show and hide image view
   const [commentUser, setCommentUser] = useState("");
-
+  const [commentsLength, setCommentsLength] = useState(0)
   let likeAmt;
   let commentAmt;
   const submitComment = props.submitComment;
@@ -35,18 +35,17 @@ export const MemoryTile = (props) => {
       })
     }
   }
-  const [comments, setComments] = useState([])
-  const getComments = async() => {
-    const endpoint = `comments?memory_uuid=${m.uuid}`
-    await handleGet(endpoint, setComments)
-  }
+  
   const getImage = async() => {
     const endpoint = `images?memory_uuid=${m.uuid}`
     await handleGet(endpoint, setImage)
   }
-
+  const getCommentsCount = async() => {
+    const endpoint = `commentCount?memory_uuid=${m.uuid}`
+    await handleGet(endpoint, setCommentsLength)
+  }
   useEffect(() => {
-    getComments();
+    getCommentsCount()
     getImage();
   }, [])
 
@@ -69,18 +68,18 @@ export const MemoryTile = (props) => {
   }
 
   //do the same client-side formatting for comments length
-  if(comments.length === 0) {
+  if(commentsLength === 0) {
     commentAmt = '';
     commentAmt = 0;
-  } else if(comments.length % 1000 === 0) {
+  } else if(commentsLength % 1000 === 0) {
     commentAmt = ''
-    commentAmt = `${comments.length / 100}k`
-  } else if(comments.length > 1000) {
+    commentAmt = `${commentsLength / 100}k`
+  } else if(commentsLength > 1000) {
     commentAmt = '';
-    commentAmt = `${Number(comments.length).toFixed(2)}k`
+    commentAmt = `${Number(commentsLength).toFixed(2)}k`
   } else {
     commentAmt = '';
-    commentAmt = comments.length;
+    commentAmt = commentsLength;
   }
   const user = sessionStorage.getItem("user_uuid")
 
@@ -98,7 +97,7 @@ export const MemoryTile = (props) => {
           setShowComments(!showComments);
           setShowImage(false);
         }}>{showComments ? `Hide ${String.fromCharCode(8593)}` : `View Comments (${commentAmt}) ${String.fromCharCode(8595)}`}</button>
-        {showComments ? <CommentsView m = {m} commentUser={commentUser} setCommentUser={setCommentUser} getComments={getComments} comments={comments} submitComment={submitComment}/> : null}
+        {showComments ? <CommentsView m = {m} commentUser={commentUser} setCommentUser={setCommentUser} submitComment={submitComment}/> : null}
 
         <button className='interaction-btn' onClick ={() => {
           setShowImage(!showImage);
@@ -119,7 +118,7 @@ export const MemoryTile = (props) => {
           checkForUser();
           setShowComments(!showComments);
         }}>{showComments ? `Hide ${String.fromCharCode(8593)}` : `View Comments (${commentAmt}) ${String.fromCharCode(8595)}`}</button>
-        {showComments ? <CommentsView m = {m} commentUser={commentUser} setCommentUser={setCommentUser} getComments={getComments} comments={comments} submitComment={submitComment}/> : null}
+        {showComments ? <CommentsView m = {m} commentUser={commentUser} setCommentUser={setCommentUser} submitComment={submitComment}/> : null}
       </section>
     )
   }
@@ -127,107 +126,4 @@ export const MemoryTile = (props) => {
   
 }
 
-//child component for each memory's comments
-const CommentsView = (props) => {
-  const recaptchaRef = useRef();
-  const {m, commentUser, setCommentUser, submitComment, comments, getComments} = props;
-
-  const user = sessionStorage.getItem("user_uuid")
-  const SITE_KEY = '6LffBlMpAAAAADK37hlL29ERh8ba5EMhRtPCli6o'
-  const recaptchaId1 = uuidv4();
-  const recaptchaId2 = uuidv4();
-  const [recaptchaValue, setRecaptchaValue] = useState(null);
-  const [newComment, setNewComment] = useState({
-    memory_uuid: m.uuid,
-    user_uuid: user,
-    comment_text: ""
-  })
-  const handleSubmit = async() => {
-    //if they hit submit but haven't entered a comment, don't do anything
-    if(newComment.comment_text === "") { 
-      return;
-    } else {
-      const token = recaptchaValue;
-      const recaptchaEndpoint = `recaptcha`
-      const recaptchaBody = {
-        recaptcha_token: token,
-      }
-      const recaptchaResponse = await handlePost(recaptchaEndpoint, recaptchaBody)
-      const recaptchaData = await recaptchaResponse.data;
-      if(recaptchaData.human) {
-        //ensures memory_uuid and user_uuid are sent in each payload
-        newComment.memory_uuid = m.uuid;
-        newComment.user_uuid = user
-        submitComment(newComment, commentUser, setNewComment, getComments, setRecaptchaValue)
-        //reset recaptcha using its id
-        recaptchaRef.current.reset(recaptchaId1)
-        recaptchaRef.current.reset(recaptchaId2);
-        
-      } else {
-        alert("You are a bot!")
-      }
-      
-    }
-    
-  }
-  const handleChange = (name, value) => {
-    setNewComment({...newComment, [name]:value})
-  }
-  const handleUserChange = (value) => {
-    setCommentUser(value)
-  }
-  if(comments.length === 0) {
-    return (
-      <section className='comments-container'>
-        <section className='no-comments-container'>
-          <h4>No comments yet, be the first to comment!</h4>
-        </section>
-        <section className='new-comment'>
-          <span>Display name: <input type="text" value={commentUser} name="commentUser" onChange={(e) => handleUserChange(e.target.value)} /></span>
-          <span>Comment: <input type="text" name="comment_text" value={newComment.comment_text} onChange={(e) => handleChange(e.target.name, e.target.value) }/></span>
-          <ReCAPTCHA size="compact" ref={recaptchaRef} id={recaptchaId1} sitekey={SITE_KEY} type="image" onChange={(val) => setRecaptchaValue(val)}/>
-          <button disabled={!recaptchaValue} className='interaction-btn submit-comment-btn' onClick = {handleSubmit}>Post comment</button>
-        </section>
-      </section>
-    )
-  } else {
-    return (
-      <section className='comments-container'>
-        <h4 className='comments-count'>Comments</h4>
-        <section className='comments-list-container'>
-          <ul className='comments-list'>
-            {
-              comments.map(com => {
-                return (
-                  <li key={uuidv4()}>
-                    <h4>{com.commenter_name.display_name}</h4>
-                    <p className='comment-text'>{com.comment_text}</p>
-                  </li>
-                )
-              })
-            }
-          </ul>
-        </section>
-        <section className='new-comment'>
-          <span>Display name: <input type="text" value={commentUser} name="commentUser" onChange={(e) => handleUserChange(e.target.value)} /></span>
-          <span>Comment: <input type="text" name="comment_text" value={newComment.comment_text} onChange={(e) => handleChange(e.target.name, e.target.value) }/></span>
-          <ReCAPTCHA size="compact" ref={recaptchaRef} id={recaptchaId2} sitekey={SITE_KEY} type="image" onChange={(val) => setRecaptchaValue(val)}/>
-          <button disabled={!recaptchaValue} className='interaction-btn submit-comment-btn' onClick = {handleSubmit}>Post comment</button>
-        </section>
-      </section>
-    )
-  
-  }
-}
-
-const ImageView = (props) => {
-  const image = props.image;
-  console.log("image:", image)
-  return (
-    <section className='image-view'>
-      <img src = {image.image_url} alt={image.image_caption}/>
-      <h4>{image.image_caption}</h4>
-    </section>
-  )
-}
